Partial conversion to GTK+
- Fix the easy warnings
- Replace deprecated gtk_text with gtk_entry

This is enough to get the application to compile and even to run, but
it segfaults when creating a card - probably due to position issues
with adding text. I think the original version used this code to
append text, but I don't know too much about this API so I can't
replace it as-is.

Will need some attention from GCC.