Create ChangeLog
And in the process, move it out of cardfile.c.

The log uses the standard GNU format. No idea if that's still a good
email for Cam, but if the last update was in 2013, maybe it is.

I'll add my entry to the log if and when there's a release to the
public.