Fix formatting
This kind of dirties the commit history (I should've done this first)
but it's a necessary evil; this code was in free-format before and
looked ug-LY!

Kernel style is my style of choice (though I usually brace functions
the same way as everything else... old habits die hard) so just using
their .clang-format is fine. It's all gpl2 anyway.