Change card_data to text_view
It's a bit stupid that the buffer and the view are seperate, but
whatever. It has the functionality we need, which is what's important.

Just need to take care of statdata too and the program will be at about
where it was in 2001, and thus ready for more dramatic changes.