Start on loading code
It's actually already something like 70% done. It can currently
gracefully load in every cardfile in my collection - though, I don't
have any RRG-format cardfiles to test it on, which I suppose is good
considering that part of the code is not yet implemented. The
Knowledge Base article on the file format describes an algorithm to
slurp OLE data, which is apparently non-trivial. I'll work on that
next, though I will be unable to test it until I get my hands on an
RRG cardfile.

I'm really proud of crd2json. I think that'll be the best part of this
codebase. For simple text-only cardfiles it's currently afaik the
best way to extract the data for preservation.