Add code to snarf OLE data
What a godawful file format. No plans to try to read this garbage and
actually display the image at this point. I don't even have an RRG
format cardfile to test this with.

If you do have one, please send it to me - along with a crate of
Special Brew - and I'll get this stuff debugged & working. For now,
I'm satisfied with this outline.