Rewrite cardwin
The broad strokes of the functionality are the same, and there's plenty
of shared code, but the code has really properly diverged.

At this point it's *better* than the original code, because it's able
to save data. Just gotta do something about those OLE objects and
bitmaps...

This is the minimum viable product. For basic usages of cardfile it's
good as-is. There's a little that I want to implement though.